+++ src/common/mod.rs
@@ -1,0 +1,6 @@
+//! Common lib for other modules
+
+pub mod audio_source;
+pub mod batch_response;
+pub mod options;
+pub mod stream_response;