+++ src/manage/scopes/response.rs
@@ -1,0 +1,27 @@
+//! Deepgram TODO API response types.
+
+use serde::{Deserialize, Serialize};
+
+/// Success message.
+///
+/// See the [Deepgram API Reference][api] for more info.
+///
+/// [api]: https://developers.deepgram.com/api-reference/#invitations
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[non_exhaustive]
+pub struct Message {
+    #[allow(missing_docs)]
+    pub message: String,
+}
+
+/// Scopes associated with the member.
+///
+/// See the [Deepgram API Reference][api] for more info.
+///
+/// [api]: https://developers.deepgram.com/api-reference/#scopes-get
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+#[non_exhaustive]
+pub struct Scopes {
+    #[allow(missing_docs)]
+    pub scopes: Vec<String>,
+}