+++ src/speak/mod.rs
@@ -1,0 +1,4 @@
+//! Speak module
+
+pub mod options;
+pub mod rest;